pyproject.toml
[tool.poetry]
name = "ably-test"
version = "0.1.0"
description = ""
authors = ["TaeJun Lee <tejunlee007@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
django-debug-toolbar = "^3.6.0"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.13.1"
drf-yasg = "^1.21.3"
Markdown = "^3.4.1"
django-filter = "^22.1"
black = "^22.8.0"
isort = "^5.10.1"
django-phonenumber-field = "^7.0.0"
phonenumbers = "^8.12.55"
PyJWT = "^2.4.0"
cryptography = "^38.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
